requirements.txt
streamlit>=1.30
openai>=1.0
pydantic>=2.0
PyGithub>=2.0
pyyaml
jinja2
